fix: render notification provider
